Apply Andrew's optimizations

+++ models/ipa_pytorch.py
@@ -420,7 +420,7 @@ class InvariantPointAttention(nn.Module):
         ################
         # [*, N_res, H, C_hidden]
         o = torch.matmul(
-            a, v.transpose(-2, -3).to(dtype=a.dtype)
+            a, v.transpose(-2, -3)
         ).transpose(-2, -3)
 
         # [*, N_res, H * C_hidden]
@@ -451,7 +451,7 @@ class InvariantPointAttention(nn.Module):
             z[0] = z[0].to(o_pt.device)
 
         # [*, N_res, H, C_z // 4]
-        pair_z = self.down_z(z[0]).to(dtype=a.dtype)
+        pair_z = self.down_z(z[0])
         o_pair = torch.matmul(a.transpose(-2, -3), pair_z)
 
         # [*, N_res, H * C_z // 4]
@@ -463,7 +463,7 @@ class InvariantPointAttention(nn.Module):
         s = self.linear_out(
             torch.cat(
                 o_feats, dim=-1
-            ).to(dtype=z[0].dtype)
+            )
         )
         
         return s
